Renamed joker_table.h to def_joker_gfx_table.h. Removed an unnecessary constant
