+++ include/def_joker_gfx_table.h
